vdraw_text.cpp: Onscreen text color can now be any RGB color.
The RGB colors are automatically converted from 32-bit to 15-bit or 16-bit
for onscreen purposes. They're saved in 32-bit format. (Well, technically
24-bit, since we're not using the alpha channel...)

+++ src/gens/video/vdraw_text.hpp
@@ -49,8 +49,8 @@ extern "C" {
 typedef struct
 {
 	uint8_t		style;
-	uint8_t		color;		// STYLE_COLOR_* values
-	uint32_t	dot_color;	// Actual RGB color value.
+	uint32_t	color;		// RGB color value. (32-bit)
+	uint32_t	dot_color;	// RGB color value. (Screen color depth.)
 	BOOL		double_size;
 	BOOL		transparent;
 } vdraw_style_t;
